clean up

+++ app/src/test/java/com/example/jens/kitchenconverter/UnitTest.java
@@ -30,7 +30,7 @@ public class UnitTest {
     public void init() throws Exception {
         System.out.println("Setting up ...");
         when(context.getResources().getStringArray(R.array.dimensions_array)).thenReturn(new String [] {"length", "mass", "volume"});
-        unit = new Unit("dm","length", (double) 0.1,context);
+        unit = new Unit("dm","length", 0.1,context);
     }
 
     @After
